refactor: Extract automatic groups logic into reusable composable
- Create useAutomaticGroups composable with shared data loading logic
- Remove duplicated code from both Card and Admin components
- Both components now use the same data source and filtering logic
- Ensure consistent behavior between Card and Admin views
- Eliminate code duplication while maintaining functionality

Benefits:
- Single source of truth for automatic groups data
- Easier maintenance and testing
- Consistent behavior across components
- Reduced bundle size
